We may need a bigger map

+++ src/lib_name/include/lib_name/lib_name.hpp
@@ -6,7 +6,7 @@
 namespace lib_name {
 
     struct lib {
-        int foo();
+        double foo();
     };
 
 }
